add-ดักอีเวนต์ทุกปุ่มเตรียมส่งข้อมูลให้ api

+++ app/src/main/java/com/thingnoy/thingnoy500v3/util/StringUtils.java
@@ -35,7 +35,7 @@ public class StringUtils {
     }
 
     @NonNull
-    public static String getCommaPriceWithBaht(Context context, int price) {
+    public static String getCommaPriceWithBaht(Context context, double price) {
         DecimalFormat formatter = new DecimalFormat("#,###,###");
         return formatter.format(price) + getBahtString(context);
     }
